feat(assert): add support for custom error message and function (master)
This commit adds support for providing a custom error message or a
function that returns an error message to the `assert` function. If the
`msg` parameter is a function, it will be invoked and the returned value
will be used as the error message. If the returned value is an instance
of `Error`, it will be thrown directly. Otherwise, the error message
will be appended to an "Assertion Error" and thrown.

The changes are made in the `src/assert.ts` file.